Traders
-balanced the trader
-traders can now trade infinitely i presume
-i added a whale which dumps at 3 second into the sim
-for now the max tps is 5000

+++ src/Clock.cpp
@@ -1,6 +1,6 @@
 #include "Clock.h"
 
-using Time = double;
+using Time = long long;
 
 Time Clock::now() const
 {
